export segment event enum from it's own file as a top level export otherwise jest fails
