+++ packages/insomnia-app/app/__jest__/redux-state-for-test.ts
@@ -10,6 +10,7 @@ export const reduxStateForTest = async (global: Partial<GlobalState> = {}): Prom
     activeWorkspaceId: null,
     activeActivity: ACTIVITY_HOME,
     activeSpaceId: BASE_SPACE_ID,
+    spaceSortOrder: 'modified-desc',
     isLoading: false,
     isLoggedIn: false,
     loadingRequestIds: {},
