+++ packages/insomnia-app/app/ui/redux/modules/__tests__/space.test.js
@@ -3,8 +3,9 @@ import thunk from 'redux-thunk';
 
 import { globalBeforeEach } from '../../../../__jest__/before-each';
 import { reduxStateForTest } from '../../../../__jest__/redux-state-for-test';
-import { SegmentEvent, trackEvent, trackSegmentEvent } from '../../../../common/analytics';
+import { trackEvent, trackSegmentEvent } from '../../../../common/analytics';
 import { ACTIVITY_HOME } from '../../../../common/constants';
+import { SegmentEvent } from '../../../../common/segment-event';
 import * as models from '../../../../models';
 import { BASE_SPACE_ID } from '../../../../models/space';
 import { getAndClearShowAlertMockArgs, getAndClearShowPromptMockArgs } from '../../../../test-utils';
