+++ packages/insomnia-app/app/ui/redux/modules/space.ts
@@ -1,5 +1,6 @@
-import { SegmentEvent, trackEvent, trackSegmentEvent } from '../../../common/analytics';
+import {  trackEvent, trackSegmentEvent } from '../../../common/analytics';
 import { ACTIVITY_HOME } from '../../../common/constants';
+import { SegmentEvent } from '../../../common/segment-event';
 import { strings } from '../../../common/strings';
 import * as models from '../../../models';
 import { BASE_SPACE_ID, Space } from '../../../models/space';
